# Conflicts:
#	src/main/java/es/plexus/hopes/hopesback/service/FormService.java
#	src/main/java/es/plexus/hopes/hopesback/service/events/SaveConsumer.java
#	src/main/java/es/plexus/hopes/hopesback/service/events/SaveEvent.java

+++ src/main/java/es/plexus/hopes/hopesback/service/events/SaveEvent.java
@@ -11,7 +11,7 @@ import lombok.RequiredArgsConstructor;
 public class SaveEvent {
 
     private  final String name;
-    private final Long patientId;
+    private final Integer patientId;
     private final FormDTO form;
 
 }
